Clean code in parse_section.R (#27)
* Use wrapper function for the most frequent case

- replace all calls of tidyr::separate with the
  argument settings col = "value", sep = "\\s+",
  extra = "merge", fill = "left", convert = TRUE
  with a call of the new function separate_into()

* Extend first and write and use second wrapper function

* Use only one wrapper function

Remove separate_into_value() and use separate_into()
instead. The idea was to use this function if the
arguments "extra" and "fill" were not explicitly set.
Now, set these arguments to their defaults ("warn")

Improve order of arguments

* Use new helper function skip_head()

This improves the readability. In skip_head(), use
seq_len() instead of 1:n to avoid an error if n is 0

* Add functions that have been added to dev meanwhile

and replace all tidyr::separate() calls with calls of
separate_into()