Start to refactor inp_to_files.R (#28)
* Use new helper function has_incomplete_features()

* Avoid duplication, use new check_package_and_class()

* Safe folder path in intermediate variable

* Extract and use new function write_section_if_in_list()

* Use new local helper function shape_file_path()

* Define sections and conversion in a configuration

Loop through the names of the configuration list. Each
list element represents a section. The configuration
stores for each section a list of length one. The list
element is the conversion function, named according to
the name of the output shape file that is to be created.

* Fix bug: change "section_name" to "section"

* No more need for shape_file_path(), reuse "shape_dir"

* Fix bug: string "shp" intead of variable

* Reintegrate body of write_section_if_in_list() in loop

and delete write_section_if_in_list().

Use message() instead of print() to inform about missing
section